requirements.txt
discord.py
undetected-chromedriver
selenium
spacy
nltk
pyotp
pytz
pytesseract
pillow
transformers
python-dotenv
ib_insync
openai
httpx
